feat: Rename from sugar to core module name (#101)
